created journal entry

+++ inc/pixa/pixa.h
@@ -92,7 +92,7 @@ int get_width();
 int get_height();
 double get_elapsed_time();
 double get_delta_time();
-double get_fps();
+float get_fps();
 
 void draw_pixel(int x, int y);
 // void draw_line(int x1, int y1, int x2, int y2);
